Moved functions to triggered.js
